requirements.txt
httpx
APScheduler
python-dotenv
